requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
github==1.2.7
idna==3.4
multidict==6.0.4
pycodestyle==2.10.0
python-dotenv==1.0.0
python-magic==0.4.27
requests==2.28.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
